lombok doesn't impress me
